pyproject.toml
[project]
name = "capabilities"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "pysmartthings==1.0.0",
    "pyyaml>=6.0.2",
    "orjson>=3.9.10",
]
